@@ -20,21 +20,27 @@
   },
   "config": {
     "platform": {
-      "php": "7.2.19"
+      "php": "8.4"
+    },
+    "allow-plugins": {
+      "php-http/discovery": true
     }
   },
   "require": {
-    "php": ">=7.2,<8.0-DEV",
+    "php": ">=8.4,<9.0-DEV",
     "ext-json": "*",
     "ext-mysqli": "*",
-    "slim/slim": "^3.12",
-    "twig/twig": "^2.11",
-    "slim/twig-view": "^2.5",
-    "symfony/dotenv": "^4.3",
-    "swaggest/json-schema": "^0.12.11",
-    "almasaeed2010/adminlte": "~2.4",
-    "sentry/sentry": "^2.1",
-    "php-http/curl-client": "^2.0"
+    "slim/slim": "^4.14",
+    "twig/twig": "^3.19",
+    "slim/twig-view": "^3.4",
+    "symfony/dotenv": "^7.2",
+    "swaggest/json-schema": "^0.12.43",
+    "almasaeed2010/adminlte": "3.*",
+    "sentry/sentry": "^4.10",
+    "php-http/curl-client": "^2.3",
+    "http-interop/http-factory-guzzle": "^1.1",
+    "php-di/php-di": "^7.0",
+    "twig/intl-extra": "^3.20"
   },
   "autoload": {
     "psr-4": {
@@ -43,14 +49,7 @@
     }
   },
   "require-dev": {
-    "phpunit/phpunit": "^8.2",
-    "jakub-onderka/php-parallel-lint": "^1.0",
-    "phploc/phploc": "^5.0",
-    "pdepend/pdepend": "^2.5",
-    "phpmd/phpmd": "^2.6",
-    "squizlabs/php_codesniffer": "^3.4",
-    "sebastian/phpcpd": "^4.1",
-    "theseer/phpdox": "^0.12.0",
-    "roave/security-advisories": "dev-master"
+    "roave/security-advisories": "dev-latest",
+    "php-parallel-lint/php-parallel-lint": "^1.4"
   }
 }
