@@ -2,38 +2,16 @@
 
 namespace Soarce\Application\Controllers;
 
-use Slim\Http\Request;
-use Slim\Http\Response;
-use Slim\Container;
-use Soarce\Config;
-use Soarce\Statistics\Database;
+use Psr\Http\Message\ResponseInterface as Response;
+use Psr\Http\Message\ServerRequestInterface as Request;
+use Soarce\Mvc\WebApplicationController;
 
-class DocsController
+class DocsController extends WebApplicationController
 {
-    /** @var Container */
-    protected $ci;
-
-    /**
-     * BaseController constructor.
-     * @param Container $dependcyInjectionContainer
-     */
-    public function __construct(Container $dependcyInjectionContainer)
-    {
-        $this->ci = $dependcyInjectionContainer;
-        $this->ci->view['activeMainMenu'] = 'docs';
-    }
-
-    /**
-     * @param  Request  $request
-     * @param  Response $response
-     * @param  array    $params
-     * @return Response
-     */
-    public function index(Request $request, Response $response, $params): Response
+    public function index(Request $request, Response $response, array $params): Response
     {
         $page = $params['page'] ?? 'index';
-        $this->ci->view['activeSubMenu'] = $page;
-        return $this->ci->view->render($response, 'docs/' . $page . '.twig');
+        return $this->view->render($response, 'docs/' . $page . '.twig');
     }
 
     /**
@@ -43,7 +21,6 @@ class DocsController
      */
     public function license(Request $request, Response $response): Response
     {
-        $this->ci->view['activeSubMenu'] = 'license';
-        return $this->ci->view->render($response, 'docs/license.twig', ['licenseText' => file_get_contents(__DIR__ . '/../../../LICENSE')]);
+        return $this->view->render($response, 'docs/license.twig', ['licenseText' => file_get_contents(__DIR__ . '/../../../LICENSE')]);
     }
 }
