@@ -2,40 +2,23 @@
 
 namespace Soarce\Application\Controllers;
 
-use Slim\Http\Request;
-use Slim\Http\Response;
-use Slim\Container;
-use Soarce\Config;
+use Psr\Http\Message\ResponseInterface as Response;
+use Psr\Http\Message\ServerRequestInterface as Request;
+use Soarce\Analyzer\Coverage;
+use Soarce\Mvc\WebApplicationController;
 use Soarce\Statistics\Database;
 
-class IndexController
+class IndexController extends WebApplicationController
 {
-    /** @var Container */
-    protected $ci;
-
-    /**
-     * BaseController constructor.
-     * @param Container $dependcyInjectionContainer
-     */
-    public function __construct(Container $dependcyInjectionContainer)
-    {
-        $this->ci = $dependcyInjectionContainer;
-        $this->ci->view['activeMainMenu'] = 'home';
-    }
-
-    /**
-      * @param  Request $request
-      * @param  Response $response
-      * @return Response
-      */
      public function index(Request $request, Response $response): Response
     {
-        $this->ci->view['configIsValid'] = Config::isValid(__DIR__ . '/../../../soarce.json');
-        $this->ci->view['configErrorMessage'] = Config::$validationError;
+        $viewVars = [
+#            'configIsValid' => Config::isValid(__DIR__ . '/../../../soarce.json'),
+#            'configErrorMessage' => Config::$validationError,
+            'DatabaseStatistics' => $this->container->get(Database::class)->getMysqlStats(),
+            'totalCoverage'  => $this->container->get(Coverage::class)->getTotalCoveragePercentage(),
+        ];
 
-        $dbstatistics = new Database($this->ci);
-        $this->ci->view['DatabaseStatistics'] = $dbstatistics->getMysqlStats();
-
-        return $this->ci->view->render($response, 'index/index.twig');
+        return $this->view->render($response, 'index/index.twig', $viewVars);
     }
 }
