@@ -2,50 +2,26 @@
 
 namespace Soarce\Application\Controllers;
 
-use Slim\Http\Request;
-use Slim\Http\Response;
-use Slim\Container;
+use Psr\Http\Message\ResponseInterface as Response;
+use Psr\Http\Message\ServerRequestInterface as Request;
+use Soarce\Analyzer\Request as RequestAnalyzer;
 use Soarce\Model\SequenceRequest;
+use Soarce\Mvc\WebApplicationController;
 
-class RequestController
+class RequestController extends WebApplicationController
 {
-    /** @var Container */
-    protected $ci;
-
-    /**
-     * BaseController constructor.
-     *
-     * @param Container $dependencyInjectionContainer
-     */
-    public function __construct(Container $dependencyInjectionContainer)
-    {
-        $this->ci = $dependencyInjectionContainer;
-        $this->ci->view['activeMainMenu'] = 'requests';
-    }
-
-    /**
-     * @param  Request  $request
-     * @param  Response $response
-     * @return Response
-     */
     public function index(Request $request, Response $response): Response
     {
-        return $this->ci->view->render($response, 'trace/index.twig');
+        return $this->view->render($response, 'trace/index.twig');
     }
 
-    /**
-     * @param  Request  $request
-     * @param  Response $response
-     * @return Response
-     */
     public function overview(Request $request, Response $response): Response
     {
-        $this->ci->view['activeSubMenu'] = 'overview';
+        $analyzer = $this->container->get(RequestAnalyzer::class);
+        $queryParams = $request->getQueryParams();
 
-        $analyzer = new \Soarce\Analyzer\Request($this->ci);
-
-        $applicationIds = $request->getParam('applicationId') ?? [];
-        $usecaseIds     = $request->getParam('usecaseId')     ?? [];
+        $applicationIds = $queryParams['applicationId'] ?? [];
+        $usecaseIds     = $queryParams['usecaseId']     ?? [];
 
         $viewParams = [
             'applications'   => $analyzer->getAppplications($usecaseIds),
@@ -54,23 +30,16 @@ class RequestController
             'usecaseIds'     => $usecaseIds,
             'requests'       => $analyzer->getRequestsOverview($usecaseIds, $applicationIds),
         ];
-        return $this->ci->view->render($response, 'request/overview.twig', $viewParams);
+        return $this->view->render($response, 'request/overview.twig', $viewParams);
     }
 
-    /**
-     * @param  Request  $request
-     * @param  Response $response
-     * @param  array    $params
-     * @return Response
-     */
     public function request(Request $request, Response $response, $params): Response
     {
-        $this->ci->view['activeSubMenu'] = 'overview';
+        $analyzer = $this->container->get(RequestAnalyzer::class);
+        $queryParams = $request->getQueryParams();
 
-        $analyzer = new \Soarce\Analyzer\Request($this->ci);
-
-        $applicationIds = $request->getParam('applicationId') ?? [];
-        $usecaseIds     = $request->getParam('usecaseId')     ?? [];
+        $applicationIds = $queryParams['applicationId'] ?? [];
+        $usecaseIds     = $queryParams['usecaseId']     ?? [];
         $requestId      = (int)($params['request'] ?? 0);
 
         if (0 === $requestId) {
@@ -82,20 +51,12 @@ class RequestController
             'usecaseIds'     => $usecaseIds,
             'request'        => $analyzer->getRequest($requestId),
         ];
-        return $this->ci->view->render($response, 'request/request.twig', $viewParams);
+        return $this->view->render($response, 'request/request.twig', $viewParams);
     }
 
-    /**
-     * @param  Request  $request
-     * @param  Response $response
-     * @param  array    $params
-     * @return Response
-     */
     public function sequence(Request $request, Response $response, $params): Response
     {
-        $this->ci->view['activeSubMenu'] = 'overview';
-
-        $analyzer = new \Soarce\Analyzer\Request($this->ci);
+        $analyzer = $this->container->get(RequestAnalyzer::class);
         $requestId = (int)($params['request'] ?? 0);
 
         if (0 === $requestId) {
@@ -123,6 +84,6 @@ class RequestController
             'colPositions'    => $colPositions,
         ];
 
-        return $this->ci->view->render($response, 'request/sequence.twig', $viewParams);
+        return $this->view->render($response, 'request/sequence.twig', $viewParams);
     }
 }
