@@ -2,11 +2,16 @@
 
 namespace Soarce\Analyzer;
 
+use mysqli;
 use Soarce\Control\Service;
 use Soarce\Control\Service\FileContent;
 
 class Coverage extends AbstractAnalyzer
 {
+    public function __construct(mysqli $mysqli, private Service $service)
+    {
+        parent::__construct($mysqli);
+    }
 
     /**
      * @param int[] $applications
@@ -14,14 +19,14 @@ class Coverage extends AbstractAnalyzer
      * @param int[] $requests
      * @return array
      */
-    public function getFiles($applications = [], $usecases = [], $requests = []): array
+    public function getFiles(array $applications = [], array $usecases = [], array $requests = []): array
     {
         $applicationList = $this->buildInStatementBody($applications);
         $usecaseList     = $this->buildInStatementBody($usecases);
         $requestList     = $this->buildInStatementBody($requests);
 
-        $sql = 'SELECT a.`id` as `applicationId`, a.`name` as `applicationName`, f.`id` as `fileId`, f.`filename` as `fileName`,
-                COUNT(distinct c.`line`) as `coveredLines`, f.`lines`
+        $sql = 'SELECT a.`id` as `applicationId`, a.`name` as `applicationName`, any_value(f.`id`) as `fileId`, f.`filename` as `fileName`,
+                COUNT(distinct c.`line`) as `coveredLines`, any_value(f.`lines`) as `lines`
             FROM `file`        f
             JOIN `application` a  ON a.`id`             = f.`application_id` ' . ($applicationList !== '' ? " and a.`id` in ({$applicationList}) " : '') . '
             JOIN `request`     r  ON r.`application_id` = a.`id` ' . ($usecaseList !== '' ? " and r.`usecase_id` in ({$usecaseList}) " : '') . ($requestList !== '' ? " and r.`id` in ({$requestList}) " : '') . '
@@ -46,10 +51,8 @@ class Coverage extends AbstractAnalyzer
      * @param int $fileId
      * @return FileContent
      */
-    public function getSource($fileId): FileContent
+    public function getSource(int $fileId): FileContent
     {
-        $controlService = new Service($this->container);
-
         $sql = 'SELECT a.`name` as `applicationName`, f.`filename` as `fileName`
             FROM `file`        f
             JOIN `application` a ON a.`id` = f.`application_id` 
@@ -62,18 +65,18 @@ class Coverage extends AbstractAnalyzer
         }
 
         $row = $result->fetch_assoc();
-        $actionable = $controlService->getServiceActionable($row['applicationName']);
+        $actionable = $this->service->getServiceActionable($row['applicationName']);
 
         return $actionable->getFile($row['fileName']);
     }
 
     /**
-     * @param  int   $fileId
-     * @param  int[] $usecaseIds
-     * @param  int[] $requestIds
+     * @param int $fileId
+     * @param int[]|null $usecaseIds
+     * @param int[]|null $requestIds
      * @return int[]
      */
-    public function getCoverage($fileId, $usecaseIds = null, $requestIds = null): array
+    public function getCoverage(int $fileId, ?array $usecaseIds, ?array $requestIds): array
     {
         $usecaseList = $this->buildInStatementBody($usecaseIds);
         $requestList = $this->buildInStatementBody($requestIds);
@@ -101,10 +104,10 @@ class Coverage extends AbstractAnalyzer
     }
 
     /**
-     * @param  int    $fileId
-     * @return mixed[]
+     * @param int $fileId
+     * @return array
      */
-    public function getFile($fileId): array
+    public function getFile(int $fileId): array
     {
         $sql = 'SELECT f.`id`, f.`application_id`, f.`filename`, LOWER(HEX(f.`md5`)) as `md5`
             FROM `file` f
@@ -120,11 +123,11 @@ class Coverage extends AbstractAnalyzer
     }
 
     /**
-     * @param  int   $fileId
-     * @param  int   $line
+     * @param int $fileId
+     * @param int $line
      * @return array
      */
-    public function getRequestsForLoc($fileId, $line): array
+    public function getRequestsForLoc(int $fileId, int $line): array
     {
         $sql = 'SELECT DISTINCT r.`id`, r.`request_id`
             FROM `coverage` c 
@@ -146,11 +149,11 @@ class Coverage extends AbstractAnalyzer
     }
 
     /**
-     * @param  int   $fileId
-     * @param  int   $line
+     * @param int $fileId
+     * @param int $line
      * @return array
      */
-    public function getUsecasesForLoC($fileId, $line): array
+    public function getUsecasesForLoC(int $fileId, int $line): array
     {
         $sql = 'SELECT u.`id`, u.`name`
             FROM `coverage` c 
@@ -170,5 +173,16 @@ class Coverage extends AbstractAnalyzer
             $ret[$row['id']] = $row['name'];
         }
         return $ret;
+    }
+
+    public function getTotalCoveragePercentage(): float
+    {
+        $sql = 'SELECT
+            (SELECT SUM(f.lines) from `file` f) as total_lines,
+            (SELECT COUNT(DISTINCT c.file_id, c.line) from `coverage` c WHERE c.covered = 1) as total_covered';
+
+        $result = $this->mysqli->query($sql)->fetch_assoc();
+
+        return $result['total_covered'] / $result['total_lines'];
     }
 }
